requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
cffi==1.15.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.4
django-celery-beat==2.5.0
django-debug-toolbar==4.2.0
django-rest-framework==0.1.0
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
gevent==23.7.0
greenlet==2.0.2
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.2
Pillow==10.0.0
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
rpds-py==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
zope.event==5.0
zope.interface==6.0
